backend/pyproject.toml
[project]
name = "ragnostic-backend"
version = "0.1.0"
description = "RAGnostic FastAPI backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-json-logger>=2.0.7,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0",
  "ruff>=0.6.0,<1.0.0",
  "mypy>=1.11.1,<2.0.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
